backend/auth-service/requirements.txt
flask==3.0.2
python-keycloak
python-dotenv
requests
flask-cors